Code/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pytesseract==0.3.10
openai==1.35.3
supabase==2.5.1
docxtpl==0.16.8
python-multipart==0.0.9
pydantic==2.7.4
Pillow==10.3.0
